Rename items to item to be consistent
