+++ src/Umbraco.Cms.Api.Management/Controllers/DataType/Item/ByEditorUiAliasController.cs
@@ -6,7 +6,7 @@ using Umbraco.Cms.Core.Mapping;
 using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Core.Services;
 
-namespace Umbraco.Cms.Api.Management.Controllers.DataType.Items;
+namespace Umbraco.Cms.Api.Management.Controllers.DataType.Item;
 
 [ApiVersion("1.0")]
 public class ByEditorUiAliasController : DatatypeItemControllerBase
