+++ src/Umbraco.Cms.Api.Management/Mapping/Item/ItemTypeMapDefinition.cs
@@ -2,11 +2,10 @@
 using Umbraco.Cms.Api.Management.ViewModels.Dictionary.Item;
 using Umbraco.Cms.Api.Management.ViewModels.DocumentType.Item;
 using Umbraco.Cms.Api.Management.ViewModels.Language.Item;
-using Umbraco.Cms.Api.Management.ViewModels.Media.Item;
 using Umbraco.Cms.Api.Management.ViewModels.MediaType.Item;
 using Umbraco.Cms.Api.Management.ViewModels.Member.Item;
 using Umbraco.Cms.Api.Management.ViewModels.MemberGroup.Item;
-using Umbraco.Cms.Api.Management.ViewModels.MemberType.Items;
+using Umbraco.Cms.Api.Management.ViewModels.MemberType.Item;
 using Umbraco.Cms.Api.Management.ViewModels.RelationType.Item;
 using Umbraco.Cms.Api.Management.ViewModels.Template.Item;
 using Umbraco.Cms.Api.Management.ViewModels.User.Item;
@@ -16,7 +15,7 @@ using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Core.Models.Entities;
 using Umbraco.Cms.Core.Models.Membership;
 
-namespace Umbraco.Cms.Api.Management.Mapping.Items;
+namespace Umbraco.Cms.Api.Management.Mapping.Item;
 
 public class ItemTypeMapDefinition : IMapDefinition
 {
