+++ src/Umbraco.Cms.Api.Management/ViewModels/Media/Item/MediaItemResponseModel.cs
