+++ src/Umbraco.Cms.Api.Management/ViewModels/Media/Item/MediaTreeItemResponseModel.cs
