+++ src/Umbraco.Cms.Api.Management/ViewModels/MemberType/Item/MemberTypeItemResponseModel.cs
@@ -1,6 +1,6 @@
 ﻿using Umbraco.Cms.Api.Management.ViewModels.Item;
 
-namespace Umbraco.Cms.Api.Management.ViewModels.MemberType.Items;
+namespace Umbraco.Cms.Api.Management.ViewModels.MemberType.Item;
 
 public class MemberTypeItemResponseModel : NamedItemResponseModelBase
 {
